META-ONLY: btrfs: warn on not-uptodate eb EIO
When btrfs reads and extent buffer, when the bios complete, we csum the
contents and mark the pages uptodate, and then the eb itself. If not all
of the pages are uptodate, or the eb is somehow not uptodate (unclear
exactly how, to me), then the read becomes an EIO. We are seeing a
steady rate of EIOs when trying to read data csums, and want to confirm
that they come from one of these paths. While this won't show what
actually happened to the bio, it will at least confirm we are on the
right track, and we can continue more confidently on debugging the bios
themselves.

Signed-off-by: Boris Burkov <boris@bur.io>