META-ONLY: btrfs: warn on bio error in endio
We suspect transient csum errors are coming from bio errors, which we
don't have good visibility into. This function is called by bio_endio
itself, not from the btrfs wq, so it is a good candidate to catch bio
errors in their proper context to try to see what is going on.

Signed-off-by: Boris Burkov <boris@bur.io>