FBONLY: btrfs: make the auto relocate error a btrfs_info
Summary:
We can fail to relocate block groups because of swapfiles or an ENOSPC,
neither of which we care about in practice, and this increases the
amount of noise we see in our alerts.  Convert this to a btrfs_info()
and print the ret so we know why we failed.  Real failures that we want
to worry about will get printed inside of relocation, this will silence
the swapfile error and anything else that relocation doesn't think is
worth complaining about.

Test Plan: testinprod

Reviewers: clm,borisb,osandov,rh0,thesweettea

Subscribers:

Tasks:

Tags: