btrfs: don't clear read-only when adding sprout device
If you follow the seed/sprout wiki, it suggests the following workflow:

btrfstune -S 1 seed_dev
mount seed_dev mnt
btrfs device add sprout_dev
mount -o remount,rw mnt

The first mount mounts the FS readonly, which results in not setting
BTRFS_FS_OPEN, and setting the readonly bit on the sb. The device add
somewhat surprisingly clears the readonly bit on the sb (though the
mount is still practically readonly, from the users perspective...).
Finally, the remount checks the readonly bit on the sb against the flag
and sees no change, so it does not run the code intended to run on
ro->rw transitions, leaving BTRFS_FS_OPEN unset.

As a result, when the cleaner_kthread runs, it sees no BTRFS_FS_OPEN and
does no work. This results in leaking deleted snapshots until we run out
of space.

I propose fixing it at the first departure from what feels reasonable:
when we clear the readonly bit on the sb during device add. I have a
reproducer of the issue here:
https://raw.githubusercontent.com/boryas/scripts/main/sh/seed/mkseed.sh
and confirm that this patch fixes it, and seems to work OK, otherwise. I
will admit that I couldn't dig up the original rationale for clearing
the bit here (it dates back to the original seed/sprout commit without
explicit explanation) so it's hard to imagine all the ramifications of
the change.

Signed-off-by: Boris Burkov <boris@bur.io>