FBONLY: btrfs: remove WARN_ON_ONCE() for the free space inode
Summary:
We have a WARN_ON_ONCE() in cow_file_range() for the free space inode,
this is because we're never supposed to COW the free space inode.
However with auto relocate on there's a window where we could have
allocated our space in the block group and then started relocating that
block group before we found the data references for that inode and thus
delete the space cache.  This means we can trip the cow codepath.  We
handle this properly by not writing out the cache at this point, so
simply remove the WARN_ON() as it's just noisy.

Test Plan: testinprod

Reviewers: clm,borisb,osandov,rh0,thesweettea

Subscribers:

Tasks:

Tags: