FBONLY: btrfs: actually throttle transactions for _end_throttle
Summary:
I changed how throttling worked by no longer committing the transaction
because it was causing problems with relocation performance.  However we
really do need to commit the transaction in these low-space cases.  With
the adjustment to the check_delayed_refs code we should be less annoying
with the relocation case and this solves the enospc problems I was
seeing with sandcastlefog.

Test Plan: Tested it on sandcastlefog with a FS that had no space and
370 pending snapshot deletes.

Reviewers: clm,borisb,osandov,rh0,thesweettea

Subscribers:

Tasks:

Tags: