fsverity: add mode sysctl
Add a "mode" sysctl to act as a killswitch for fs-verity. The motivation
is that in case there is an unforeseen performance regression or bug in
verity as one attempts to use it in production, being able to disable it
live gives a lot of peace of mind. One could also run in audit mode for
a while before switching to the enforcing mode.

The modes are:
disable: verity has no effect, even if configured or used
audit: run the verity logic, permit ioctls, but only log on failures
enforce: fully enabled; the default and the behavior before this patch

This also precipitated re-organizing sysctls for verity as previously
the only sysctl was for signatures so sysctls and signatures were
coupled.

One slightly subtle issue is what errors should "audit" swallow.
Certainly verification or signature errors, but there is a slippery
slope leading to inconsistent states with half set up verity if you try
to ignore errors in enabling verity on a file. The intent of audit mode
is to still fail normally in verity specific ioctls, but to leave the
file be from the perspective of normal filesystem APIs. However, we must
still disallow writes in audit mode, or else it would immediately lead
to invalid verity state.

Signed-off-by: Boris Burkov <boris@bur.io>