FB-ONLY: KLP,kworker: explicitly transition kworker tasks to new KLP version
Song is looking at a more generic, cleaner solution for this upstream.
However, while that longer term project is ongoing, we probably want a
short term hack to keep the number of KLP transition failures down.

The data we have in the kernel_livepatch logview data set suggests that
almost all of the KLP transition failures in our fleet are due to either
tasks stuck in autofs_wait, for which we have another workaround, or due
to kworker threads working through a queue of long lasting work items,
mostly btrfs_preempt_reclaim_metadata_space.

Add a hack to kworker to attempt transitioning to a new KLP, until
we can replace it with something better.

Signed-off-by: Rik van Riel <riel@fb.com>