btrfs: adjust the unclamp logic
Summary:
btrfs: adjust the unclamp logic

We only unclamp if we looped through the preemptive flushing once.
However we could have a relatively high clamp value, and thus loop
through a few times.

Adjust the unclamp logic to handle the cases where we're spinning our
wheels too much on the same flushing state if it isn't DELALLOC.
Additionally if we have to adjust up our reclaim amount then there isn't
much left to reclaim and we know we can unclamp in that case as well.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

Signed-off-by: Josef Bacik <josef@toxicpanda.com>