Make sure RO|EX mmap gets 2MB aligned virtual address
Summary:

Currently there are shared library text files that can be mmaped using Transparent Hugepages (THPs) but are not
due to the VMA not being aligned to 2MB.

This fixes the alignment for VMAs if CONFIG_TRANSPARENT_HUGEPAGE is enabled and the region is of length >= 2MB. This takes
care of aligning the virtual address such that THPs can be used to mmap the file.

Test Plan:

Include in v5.12 builds. WhatsApp msgdb team would like to test it out in their test tier as well.

Reviewers:

Subscribers:

Tasks:

T123560013

Tags:

Differential Revision: https://phabricator.intern.facebook.com/D37965842